% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{cr_collect}
\alias{cr_collect}
\title{Collect crawl results}
\usage{
cr_collect(crawler)
}
\arguments{
\item{crawler}{A \link{Crawler} that has been run.}
}
\value{
A tibble of all records pushed by handlers.
}
\description{
Collect crawl results
}
