% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawlee-package.R
\docType{package}
\name{crawlee-package}
\alias{crawlee}
\alias{crawlee-package}
\title{crawlee: A Tidy R Interface for Reproducible Web Crawling}
\description{
A tidy, pipe-friendly toolkit for reproducible web crawling and structured data collection, inspired by the architecture of the 'Crawlee' library. Provides a unified crawler with a deduplicating, resumable request queue, content-type aware handlers, structured storage backends and rich console logging via 'cli'. Designed for the collection and analysis of public and governmental data from HTML pages, sitemaps, RSS feeds and PDF documents.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/StrategicProjects/crawlee}
  \item \url{https://strategicprojects.github.io/crawlee/}
  \item Report bugs at \url{https://github.com/StrategicProjects/crawlee/issues}
}

}
\author{
\strong{Maintainer}: André Leite \email{leite@de.ufpe.br}

Authors:
\itemize{
  \item André Leite \email{leite@de.ufpe.br}
}

}
\keyword{internal}
